\name{chp.11.9.1}
\alias{chp.11.9.1}

\title{Breast Cancer, 43 countries}

\description{
  World Health Organization data years 1920-2000.
    Federico Girosi and Gary King, 2006, "Cause of Death Data",
  hdl:1902.1/UOVMCPSWOL UNF:3:9JU+SmVyHgwRhAKclQ85Cg==Murray Research Archive  
}

\usage{data(chp.11.9.1)}

\format{
  A YourCast 'dataobj' input object: a list of a list of cross sectional
  dataframes, dataframe of geographic names, and string indicationg CSID
  coding rule. Used to calculate empirical priors for the 'map' model
  used in figure 11.9 in \emph{Demographic Forecasting}.}

\source{World Health Organization}

\references{
  \url{http://gking.harvard.edu/yourcast}
}

\keyword{datasets}

\seealso{demo(chp.11.9)}
