\name{chp.2.7.1}
\alias{chp.2.7.1}

\title{Suicide deaths (male), USA}

\description{
  World Health Organization data years 1920-2000.
    Federico Girosi and Gary King, 2006, "Cause of Death Data",
  hdl:1902.1/UOVMCPSWOL UNF:3:9JU+SmVyHgwRhAKclQ85Cg==Murray Research Archive  
}

\usage{data(chp.2.7.1)}

\format{
  A YourCast 'dataobj' input object: a list of a list of cross sectional
  dataframes, dataframe of geographic names, and string indicationg CSID
  coding rule. Used to create figure 2.7 in \emph{Demographic Forecasting}.}

\source{World Health Organization}

\references{
  \url{http://gking.harvard.edu/yourcast}
}

\keyword{datasets}

\seealso{demo(chp.2.7)}
