\name{cntry.codes}
\alias{cntry.codes}

\title{Country codes and names for yourprep example}

\description{
  World Health Organization data 1920-2000.
    Federico Girosi and Gary King, 2006, "Cause of Death Data",
  hdl:1902.1/UOVMCPSWOL UNF:3:9JU+SmVyHgwRhAKclQ85Cg== Murray Research Archive  
}

\usage{data(cntry.codes)}

\details{191 countries included.}

\format{A fixed width \file{.txt} file.}

\source{WHO}

\references{\url{http://gking.harvard.edu/yourcast}}

\seealso{example(yourprep)}

\keyword{datasets}
