\name{csid204565}
\alias{csid204565}

\title{yourprep example file for Belize, age 65}

\description{
  World Health Organization data years 1920-2000.
    Federico Girosi and Gary King, 2006, \emph{Cause of Death Data},
  hdl:1902.1/UOVMCPSWOL UNF:3:9JU+SmVyHgwRhAKclQ85Cg==Murray Research Archive  
}

\usage{data(csid204565)}

\format{
  A fixed width \file{.txt} file.}

\details{Example cross section dataframe for Belize, age 65. To be used
  in with Example section in user's manual.}

\source{World Health Organization}

\references{\url{http://gking.harvard.edu/yourcast}}

\keyword{datasets}
