\name{print.summary.yourcast}
\alias{print.summary.yourcast}

\title{Print method for yourcast summary output}

\description{summary.yourcast class print function}

\usage{\method{print}{summary.yourcast}(x, ...)}

\arguments{
  \item{x}{\code{\link{summary.yourcast}} output object to be printed}
  \item{...}{Arguments to be passed to or from other methods.}
}

\references{\url{http://gking.harvard.edu/yourcast}}

\author{Jon Bischof \email{jbischof@fas.harvard.edu}}

\seealso{\code{\link{yourcast}},\code{\link{summary.yourcast}}}

