\name{print.yourcast}
\alias{print.yourcast}

\title{Print method for yourcast output}

\description{yourcast class print function}

\usage{
\method{print}{yourcast}(x, ...)}

\arguments{
  \item{x}{\code{\link{yourcast}} output object to be printed}
  \item{...}{Arguments to be passed to or from other methods.}
}

\references{\url{http://gking.harvard.edu/yourcast}}

\author{Jon Bischof \email{jbischof@fas.harvard.edu}}

\seealso{\code{\link{yourcast}}}

