\name{print.yourprep}
\alias{print.yourprep}

\title{Print method for yourprep output}

\description{yourprep class print function}

\usage{\method{print}{yourprep}(x, ...)}

\arguments{
  \item{x}{\code{\link{yourprep}} output object to be printed}
  \item{...}{Arguments to be passed to or from other methods.}
}

\references{\url{http://gking.harvard.edu/yourcast}}

\author{Jon Bischof \email{jbischof@fas.harvard.edu}}

\seealso{\code{\link{yourprep}}}

