\name{proximity}
\alias{proximity}

\title{Proximity data for yourprep example}

\description{
  World Health Organization data years 1920-2000.
    Federico Girosi and Gary King, 2006, "Cause of Death Data",
  hdl:1902.1/UOVMCPSWOL UNF:3:9JU+SmVyHgwRhAKclQ85Cg== Murray Research Archive  
}

\usage{data(proximity)}

\format{
  A fixed width \file{.txt} file.}

\source{WHO}

\references{\url{http://gking.harvard.edu/yourcast}}

\seealso{example(yourprep)}

\keyword{datasets}




