\name{user.prompt}
\alias{user.prompt}

\title{Scan user input}

\description{
  Gets user response to continue with the demos simulation or to quit.
}

\usage{user.prompt()}

\details{Allows users to terminate demos early.}

\value{None.}

\references{\url{http://gking.harvard.edu/yourcast}}

\author{Jon Bischof \email{jbischof@fas.harvard.edu}}

\seealso{See demos; for example: demo(chp.11.4)}

